updated doctor rating to double, recalculate and update rating on new review

+++ src/main/java/com/Medicare/model/Doctor.java
@@ -30,7 +30,7 @@ public class Doctor {
     private String specialty;
     private String specialityDetails;
     private float Fees;
-    private Integer Rating;
+    private double Rating;
     private String Bio;
     private LocalTime startTime;
     private LocalTime endTime;
